+++ Changelog.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Se agrego el modelo para el jefe (20/08/2024)
 - Se creo el script para el jefe (20/08/2024)
 - Se agrego material para hacer que los proyectiles tengan efecto de brillo (20/08/2024)
+- Se agregaron etiquetas para naves y proyectiles para asi ignorar colisiones entre proyectiles (20/08/2024)
 
 ### Changed